adds routeHandlers folder, moves weather functionality to weather.js
