+++ server.js
@@ -9,6 +9,7 @@ const cors = require('cors');
 const weatherData = require('./data/weather.json');
 const app = express(); // gives back return value
 app.use(cors()); // acts as middleware - intermediary between server and requests
+const handleGetWeather = require('./routeHandlers/weather');
 
 // open route for weather
 app.get('/weather', handleGetWeather);
@@ -20,21 +21,7 @@ app.get('/*', (req, res) => res.status(404).send('Route Not Found'));
 app.listen(process.env.PORT, () => console.log('server is listening on PORT 3001'));
 
 
-async function handleGetWeather(req, res) {
 
-  console.log(req.query);
-
-  try {
-    const liveWeather = await axios.get(`http://api.weatherbit.io/v2.0/forecast/daily?lat=${req.query.lat}&lon=${req.query.lon}&key=${process.env.WEATHER_API_KEY}&units=I&days=7`);
-    console.log(liveWeather.data);
-    const liveWeatherRes = liveWeather.data.data.map(day => new Forecast(day));
-    // send city weather data description to client
-    res.status(200).send(liveWeatherRes);
-  } catch (error) {
-
-    res.status(400).send(`There was an error retrieving weather data for ${req.query.city_name}.`);
-  }
-}
 
 async function handleGetMovie(req, res) {
   try {
@@ -45,14 +32,6 @@ async function handleGetMovie(req, res) {
     res.status(200).send(movieRes);
   } catch (error) {
     res.status(500).send(`There was server error retrieving movie data for ${req.query.city_name}.`);
-  }
-}
-
-class Forecast {
-  constructor(obj) {
-    this.date = obj.datetime;
-    this.description = `High of ${obj.max_temp}, Low of ${obj.low_temp}, with ${obj.weather.description} `;
-
   }
 }
